pyproject.toml
[project]
name = "emergent"
version = "0.1.0"
description = "Autonomous Agent Runtime — personal, local-first, Telegram-accessible"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0",
    "anthropic>=0.82.0",
    "apscheduler>=3.0",
    "chromadb>=1.5.0",
    "httpx>=0.28.1",
    "psutil>=7.2.2",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]

[project.scripts]
emergent = "emergent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/emergent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests that call real LLM (budget-capped)",
    "security: red team / security tests",
    "expensive: tests with significant API cost",
]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
    "types-psutil>=7.2.2.20260130",
]
